Load wehoop women's college basketball player box scores

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

load_wbb_player_box(
  seasons = most_recent_wbb_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons
A vector of 4-digit years associated with given women's college basketball seasons. (Min: 2006)

...
Additional arguments passed to an underlying function that writes the season data into a database (used by update_wbb_db()).

dbConnection
A DBIConnection object, as returned by DBI::dbConnect()

tablename
The name of the play by play data table within the database

Examples

# \donttest{
load_wbb_player_box(2021)

#> # A tibble: 75,026 × 34
#>    athlete_display_… team_short_displ… min   fg    fg3   ft    oreb  dreb  reb  
#>    <chr>             <chr>             <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 Kristin Scott     Iowa State        28    5-11  1-3   4-4   1     4     5    
#>  2 Kylie Feuerbach   Iowa State        19    2-6   0-2   4-5   1     2     3    
#>  3 Emily Ryan        Iowa State        20    1-4   1-3   0-0   0     6     6    
#>  4 Lexi Donarski     Iowa State        23    2-10  0-2   0-0   0     1     1    
#>  5 Ashley Joens      Iowa State        36    10-15 3-7   10-10 5     6     11   
#>  6 Morgan Kane       Iowa State        9     4-4   0-0   0-1   3     1     4    
#>  7 Aubrey Joens      Iowa State        25    4-14  4-12  0-0   1     3     4    
#>  8 Maddie Frederick  Iowa State        5     0-1   0-1   0-0   0     1     1    
#>  9 Rae Johnson       Iowa State        15    1-4   1-4   0-0   0     0     0    
#> 10 Madison Wise      Iowa State        20    2-2   1-1   0-0   0     7     7    
#> # … with 75,016 more rows, and 25 more variables: ast <chr>, stl <chr>,
#> #   blk <chr>, to <chr>, pf <chr>, pts <chr>, starter <lgl>, ejected <lgl>,
#> #   did_not_play <lgl>, active <lgl>, athlete_jersey <chr>, athlete_id <chr>,
#> #   athlete_short_name <chr>, athlete_headshot_href <chr>,
#> #   athlete_position_name <chr>, athlete_position_abbreviation <chr>,
#> #   team_name <chr>, team_logo <chr>, team_id <chr>, team_abbreviation <chr>,
#> #   team_color <chr>, game_id <int>, season <int>, season_type <int>, …

# }