Load wehoop women's college basketball schedule

helper that loads multiple seasons from the data repo either into memory or writes it into a db using some forwarded arguments in the dots

Usage

load_wbb_schedule(
  seasons = most_recent_wbb_season(),
  ...,
  dbConnection = NULL,
  tablename = NULL
)

Arguments

seasons

A vector of 4-digit years associated with given women's college basketball seasons. (Min: 2002)

...

Additional arguments passed to an underlying function that writes the season data into a database (used by update_wbb_db()).

dbConnection

A DBIConnection object, as returned by DBI::dbConnect()

tablename

The name of the play by play data table within the database

Value

Returns a tibble

Examples

# \donttest{
  try(load_wbb_schedule())
#> ──────────────────────────────────────────────────────────────── wehoop 2.1.0 ──
#> # A tibble: 5,923 × 84
#>         id uid   date  attendance time_valid neutral_site conference_competition
#>      <int> <chr> <chr>      <dbl> <lgl>      <lgl>        <lgl>                 
#>  1  4.02e8 s:40… 2024…      18300 TRUE       TRUE         FALSE                 
#>  2  4.02e8 s:40… 2024…       1545 TRUE       TRUE         FALSE                 
#>  3  4.02e8 s:40… 2024…      18284 TRUE       TRUE         FALSE                 
#>  4  4.02e8 s:40… 2024…          0 TRUE       TRUE         FALSE                 
#>  5  4.02e8 s:40… 2024…          0 TRUE       TRUE         FALSE                 
#>  6  4.02e8 s:40… 2024…       4333 TRUE       FALSE        FALSE                 
#>  7  4.02e8 s:40… 2024…       1694 TRUE       FALSE        FALSE                 
#>  8  4.02e8 s:40… 2024…      10869 TRUE       TRUE         FALSE                 
#>  9  4.02e8 s:40… 2024…       3710 TRUE       FALSE        FALSE                 
#> 10  4.02e8 s:40… 2024…       2198 TRUE       FALSE        FALSE                 
#> # ℹ 5,913 more rows
#> # ℹ 77 more variables: play_by_play_available <lgl>, recent <lgl>,
#> #   start_date <chr>, notes_type <chr>, notes_headline <chr>,
#> #   broadcast_market <chr>, broadcast_name <chr>, type_id <int>,
#> #   type_abbreviation <chr>, venue_id <int>, venue_full_name <chr>,
#> #   venue_address_city <chr>, venue_address_state <chr>, venue_indoor <lgl>,
#> #   status_clock <dbl>, status_display_clock <chr>, status_period <dbl>, …
# }