Get Women's College Basketball NET Rankings for the current date from the NCAA website

Get Women's College Basketball NET Rankings for the current date from the NCAA website

wbb_ncaa_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
wbb_ncaa_NET_rankings()

#> # A tibble: 343 x 9
#>     Rank Previous School       Conference Record Road  Neutral Home  `Non Div I`
#>    <int>    <int> <chr>        <chr>      <chr>  <chr> <chr>   <chr> <chr>      
#>  1     1        2 Stanford     Pac-12     21-2   13-1  3-1     5-0   0-0        
#>  2     2        1 UConn        Big East   20-1   10-1  0-0     10-0  0-0        
#>  3     3        3 South Carol~ SEC        19-3   7-2   2-0     10-1  0-0        
#>  4     4        4 Baylor       Big 12     19-2   8-1   0-0     11-1  0-0        
#>  5     5        5 Maryland     Big Ten    18-2   7-1   2-1     9-0   0-0        
#>  6     6        6 Louisville   ACC        20-2   6-1   1-0     13-1  0-0        
#>  7     7        9 UCLA         Pac-12     14-4   5-2   1-0     8-2   0-0        
#>  8     8        7 Oregon       Pac-12     13-6   6-2   0-1     7-3   0-0        
#>  9     9        8 NC State     ACC        16-2   5-2   0-0     11-0  0-0        
#> 10    10       10 Indiana      Big Ten    16-4   8-2   0-0     8-2   0-0        
#> # ... with 333 more rows