Get Women's college basketball NET rankings for the current date from the NCAA website

Get Women's college basketball NET rankings for the current date from the NCAA website

ncaa_wbb_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
ncaa_wbb_NET_rankings()

#> # A tibble: 343 x 9
#>     rank previous school         conference record road  neutral home  non_div_i
#>    <int>    <int> <chr>          <chr>      <chr>  <chr> <chr>   <chr> <chr>    
#>  1     1        1 Stanford       Pac-12     31-2   13-1  12-1    6-0   0-0      
#>  2     2        2 UConn          Big East   28-2   10-1  7-1     11-0  0-0      
#>  3     3        3 Baylor         Big 12     28-3   10-1  6-1     12-1  0-0      
#>  4     4        5 Maryland       Big Ten    26-3   9-1   7-2     10-0  0-0      
#>  5     5        4 South Carolina SEC        26-5   7-3   9-1     10-1  0-0      
#>  6     6        6 Louisville     ACC        26-4   7-1   6-2     13-1  0-0      
#>  7     7       15 Arizona        Pac-12     21-6   4-3   6-2     11-1  0-0      
#>  8     8        8 UCLA           Pac-12     17-6   5-2   4-2     8-2   0-0      
#>  9     9        9 Indiana        Big Ten    21-6   8-2   3-2     10-2  0-0      
#> 10    10       10 Oregon         Pac-12     15-9   6-2   2-3     7-4   0-0      
#> # … with 333 more rows