Get Women's college basketball NET rankings for the current date from the NCAA website

Get Women's college basketball NET rankings for the current date from the NCAA website

ncaa_wbb_NET_rankings()

Value

Returns a tibble

Author

Saiem Gilani

Examples

# Get current NCAA NET rankings
# \donttest{
  try(ncaa_wbb_NET_rankings())
#> # A tibble: 356 × 9
#>     rank previous school         conference record road  neutral home  non_div_i
#>    <int>    <int> <chr>          <chr>      <chr>  <chr> <chr>   <chr> <chr>    
#>  1     1        1 South Carolina SEC        12-0   2-0   4-0     6-0   0-0      
#>  2     2        2 NC State       ACC        11-2   2-0   2-0     7-2   0-0      
#>  3     3        3 North Carolina ACC        11-0   4-0   2-0     5-0   0-0      
#>  4     4        4 Texas          Big 12     9-1    3-1   1-0     5-0   0-0      
#>  5     5        5 Stanford       Pac-12     8-3    2-1   2-1     4-1   0-0      
#>  6     6        6 Louisville     ACC        10-1   3-0   1-1     6-0   0-0      
#>  7     7        7 Nebraska       Big Ten    12-0   3-0   1-0     8-0   0-0      
#>  8     8        8 Arizona        Pac-12     10-0   1-0   4-0     5-0   0-0      
#>  9     9        9 Iowa St.       Big 12     11-1   1-1   3-0     7-0   0-0      
#> 10    10       14 Indiana        Big Ten    10-2   3-0   1-1     6-1   0-0      
#> # … with 346 more rows
# }